Update default.xml
Due to DMCA notification mail.  Svox repo was disabled.
Hello,

We received a DMCA takedown notice regarding your repository xhteam/external-svox.
As such, we have disabled public access to the repository.

The notice has been publicly posted at:

https://github.com/github/dmca/blob/master/2014-12-28-Cambridge-Mobile.md

If you believe that your repository was disabled as a result of mistake or misidentification, you have the right to file a counter notice and have the repository reinstated. Our help articles provide more details about how the DMCA notice-and-takedown process works at GitHub and how to file a counter notice with us if you choose:

https://help.github.com/articles/dmca-takedown-policy
https://help.github.com/articles/guide-to-submitting-a-dmca-counter-notice

If you have any questions about the process or the risks in filing a counter notice, we suggest that you consult with a lawyer.

Thanks,
The GitHub Team